Apply more build bug fixes from Tonko Juricic.

+++ libcore/SWFMatrix.cpp
@@ -309,7 +309,8 @@ SWFMatrix::get_y_scale() const
 double
 SWFMatrix::get_rotation() const
 {
-    return atan2(shx, sx); // more successes in misc-ming.all/SWFMatrix_test.c
+    // more successes in misc-ming.all/SWFMatrix_test.c
+    return atan2(static_cast<double>(shx), sx); 
 }
 
 // private
